+++ dev4press/core/ui/admin/Panel.php
@@ -3,7 +3,7 @@
 namespace Dev4Press\v50\Core\UI\Admin;
 
 use Dev4Press\v50\Core\Quick\KSES;
-use Dev4Press\v50\WordPress\Admin\Table;
+use Dev4Press\v50\Library;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
@@ -207,7 +207,7 @@ abstract class Panel {
 	}
 
 	public function forms_path_library() : string {
-		return $this->a()->path . 'vendor/dev4press-library/forms/';
+		return $this->a()->path . Library::instance()->base_path() . '/forms/';
 	}
 
 	public function forms_path_plugin() : string {