+++ forms/sidebar-dashboard.php
@@ -1,6 +1,7 @@
 <?php
 
 use Dev4Press\v50\Core\Quick\KSES;
+use Dev4Press\v50\Library;
 use function Dev4Press\v50\Functions\panel;
 
 if ( ! defined( 'ABSPATH' ) ) {
@@ -34,7 +35,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 
         <div class="d4p-links-group buy-me-a-coffee">
             <a href="https://www.buymeacoffee.com/millan" target="_blank" rel="noopener">
-                <img alt="BuyMeACoffee" src="<?php echo esc_url( panel()->a()->url . 'vendor/dev4press-library/resources/gfx/buy_me_a_coffee.png' ); ?>"/>
+                <img alt="BuyMeACoffee" src="<?php echo esc_url( panel()->a()->url . Library::instance()->base_path() . '/resources/gfx/buy_me_a_coffee.png' ); ?>"/>
             </a>
         </div>